fastapi-backend/pyproject.toml
[project]
name = "fastapi-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosmtplib>=5.0.0",
    "fastapi>=0.128.0",
    "fastapi-mail>=1.6.1",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "slowapi>=0.1.9",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.40.0",
]
